Add a pyproject.toml + workers on http server
