+++ .gitignore
@@ -2,3 +2,4 @@
 tmp*
 *.log
 __pycache__
+*.egg-info
