+++ requirements.txt
@@ -1,0 +1,4 @@
+websockets==15.0.1
+fastapi==0.116.0
+uvicorn==0.35.0
+python-dotenv==1.1.1
